Restructure resolver package
