+++ internal/pipeline/resolver/config_test.go
@@ -1,4 +1,4 @@
-package pipeline
+package resolver
 
 import (
 	"os"
