+++ internal/pipeline/resolver/path_test.go
@@ -1,4 +1,4 @@
-package pipeline
+package resolver
 
 import (
 	"net/http"
